feat: Updated env utils
